🐛 멤버가 탈퇴할 때 참여하고 있는 프로젝트가 유지되는 버그 수정

+++ src/main/java/knu/team1/be/boost/member/service/MemberService.java
@@ -9,6 +9,8 @@ import knu.team1.be.boost.member.dto.MemberNotificationResponseDto;
 import knu.team1.be.boost.member.dto.MemberResponseDto;
 import knu.team1.be.boost.member.entity.Member;
 import knu.team1.be.boost.member.repository.MemberRepository;
+import knu.team1.be.boost.projectMembership.entity.ProjectRole;
+import knu.team1.be.boost.projectMembership.repository.ProjectMembershipRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -19,6 +21,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final ProjectMembershipRepository projectMembershipRepository;
 
     public MemberResponseDto getMember(UUID memberId) {
         Member member = memberRepository.findById(memberId)
@@ -66,6 +69,21 @@ public class MemberService {
                 ErrorCode.MEMBER_NOT_FOUND,
                 "memberId: " + memberId
             ));
+
+        boolean hasOwnedProjects = projectMembershipRepository.existsByMemberIdAndRole(
+            memberId,
+            ProjectRole.OWNER
+        );
+
+        if (hasOwnedProjects) {
+            throw new BusinessException(
+                ErrorCode.MEMBER_HAS_OWNED_PROJECTS,
+                "memberId: " + memberId
+            );
+        }
+
+        projectMembershipRepository.softDeleteAllByMemberId(memberId);
+
         memberRepository.delete(member);
     }
